# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alumno",
        "description": "Documentación para la API REST del sistema de Gestion de Alumnos",
        "contact": {
            "email": "neraricardo2013@gmail.io"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/sedes/{id_sede}/alumnos/{id_alumno}": {
            "get": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Mostrar alumno",
                "description": "Recupera al alumno de acuerdo al id y sede",
                "operationId": "show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_sede"
                    },
                    {
                        "$ref": "#/components/parameters/id_alumno"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo alumno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumno"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Editar alumno",
                "description": "Edita el alumno de acuerdo al id y sede",
                "operationId": "update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_sede"
                    },
                    {
                        "$ref": "#/components/parameters/id_alumno"
                    }
                ],
                "requestBody": {
                    "description": "Datos del alumno",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Alumno"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve un solo alumno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumno"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error Validator."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Eliminar alumno",
                "description": "Elimina al alumno de acuerdo al id. Si el alumno posse inscripciones activas, no puede ser eliminado",
                "operationId": "destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_sede"
                    },
                    {
                        "$ref": "#/components/parameters/id_alumno"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo alumno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alumno"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Si el alumno posse inscripciones activas, no puede ser eliminado."
                    }
                }
            }
        },
...
        "/becas/{id_beca}": {
            "get": {
                "tags": [
                    "Becas"
                ],
                "summary": "Mostrar beca",
                "description": "Recupera la beca de acuerdo al id",
                "operationId": "show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_beca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo beca.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Beca"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Becas"
                ],
                "summary": "Editar beca",
                "description": "Edita la beca de acuerdo al id",
                "operationId": "update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_beca"
                    }
                ],
                "requestBody": {
                    "description": "Datos del beca",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Beca"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve una sola beca.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Beca"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error Validator."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Becas"
                ],
                "summary": "Eliminar beca",
                "description": "Elimina la beca de acuerdo al id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_beca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo una beca.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Beca"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/carreras/{id_carrera}": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Mostrar carreras",
                "description": "Recupera la carrera de acuerdo al id",
                "operationId": "show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_carrera"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo carreras.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Editar carreras",
                "description": "Edita la carrera de acuerdo al id",
                "operationId": "update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_carrera"
                    }
                ],
                "requestBody": {
                    "description": "Datos del carrera",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carrera"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve una sola carrera.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error Validator."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Eliminar carrera",
                "description": "Elimina la carreras de acuerdo al id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_carrera"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo una carrera.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Carrera"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/departamentos/{id_departamento}": {
            "get": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Mostrar departamento",
                "description": "Recupera el departamento de acuerdo al id",
                "operationId": "show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_departamento"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo departamento.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Departamento"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Editar departamento",
                "description": "Edita el departamento de acuerdo al id",
                "operationId": "update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_departamento"
                    }
                ],
                "requestBody": {
                    "description": "Datos del departamento",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Departamento"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve un solo departamento.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Departamento"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error Validator."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Eliminar departamento",
                "description": "Elimina al departamento de acuerdo al id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id_departamento"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve un solo departamento.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Departamento"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sedes": {
            "get": {
                "tags": [
                    "Sedes"
                ],
                "summary": "Listado de sedes",
                "description": "Mostrar todas las sedes",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "listado de todas las sedes."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Sedes"
                ],
                "summary": "Nueva sede",
                "description": "Guardar nueva sede",
                "operationId": "create",
                "requestBody": {
                    "description": "Datos para crear nueva sede",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Sede"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve la sede creada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sede"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error Validator."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Beca": {
                "required": [
                    "nombre",
                    "porcentaje"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "id_usuario": {
                        "description": "Usuario responsable del registro",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true
                    },
                    "nombre": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "descripcion": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "porcentaje": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "id_sede": {
                "name": "id_sede",
                "in": "path",
                "description": "Identificación de la sede",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "id_alumno": {
                "name": "id_alumno",
                "in": "path",
                "description": "Identificación del alumno",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "id_departamento": {
                "name": "id_departamento",
                "in": "path",
                "description": "Identificación del departamento",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "id_beca": {
                "name": "id_beca",
                "in": "path",
                "description": "Identificación de la beca",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "id_carrera": {
                "name": "id_carrera",
                "in": "path",
                "description": "Identificación de la carrera",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        }
    },
...
}